+++ pentobi/i18n/qml_en.ts
@@ -1348,5 +1348,14 @@
         <source>Too many arguments</source>
         <translation type="unfinished"></translation>
     </message>
+    <message>
+        <source>Use layout optimized for desktop.</source>
+        <extracomment>Description for command line option --desktop</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>--mobile and --desktop cannot be used both</source>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 </TS>